Fix/task 2

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -33,6 +34,22 @@ public class ImageController {
         List<Image> images = imageService.getAllImages();
         model.addAttribute("images", images);
         return "images";
+    }
+
+    /*
+        This method has two params user,session which verifies the user object with
+        the userLoggedIn
+        Return true if both are Id's are matched.
+        Return false if Id's does not match.
+     */
+
+    private Boolean verifyLoggedInUser(User user, HttpSession session) {
+        User userLoggedIn = (User) session.getAttribute("loggeduser");
+        if (user.getId() == userLoggedIn.getId()) {
+            return true;
+        } else {
+            return false;
+        }
     }
 
     //This method is called when the details of the specific image with corresponding title are to be displayed
@@ -92,13 +109,22 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
         Image image = imageService.getImage(imageId);
+        User currentUser  = image.getUser();
+        if(verifyLoggedInUser(currentUser, session)) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+            String imageTitle = imageService.getImage(imageId).getTitle();
+            String error = "Only the owner of the image can delete the image";
+            redirectAttributes.addAttribute("deleteError", error).addFlashAttribute("deleteError", error); // add deleteError attributes with data before direction
+            return "redirect:/images/" + imageId + '/' + imageTitle;
+        }
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +166,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirectAttributes,Model model) {
+
+        Image image = imageService.getImage(imageId);
+        User currentUser = image.getUser();
+        if(verifyLoggedInUser(currentUser, session)) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            String imageTitle = imageService.getImage(imageId).getTitle();
+            String error = "Only the owner of the image can delete the image";
+            redirectAttributes.addAttribute("deleteError", error).addFlashAttribute("deleteError", error); // add deleteError attributes with data before direction
+            return "redirect:/images/" + imageId + '/' + imageTitle;
+        }
+
     }
